+++ Meridian59/Data/Models/ObjectBase.cs
@@ -1078,7 +1078,7 @@ namespace Meridian59.Data.Models
         /// </summary>
         /// <param name="Tick"></param>
         /// <param name="Span"></param>
-        public virtual void Tick(long Tick, long Span)
+        public virtual void Tick(double Tick, double Span)
         {
             // update animation
             Animation.Tick(Tick, Span);
