+++ Meridian59/Data/Models/TradeInfo.cs
@@ -222,7 +222,7 @@ namespace Meridian59.Data.Models
         /// </summary>
         /// <param name="Tick"></param>
         /// <param name="Span"></param>
-        public void Tick(long Tick, long Span)
+        public void Tick(double Tick, double Span)
         {
             // update items partner            
             foreach (ObjectBase obj in ItemsPartner)
