+++ Meridian59/Files/ROO/RooSector.cs
@@ -685,7 +685,7 @@ namespace Meridian59.Files.ROO
         /// </summary>
         /// <param name="Tick"></param>
         /// <param name="Span"></param>
-        public void Tick(long Tick, long Span)
+        public void Tick(double Tick, double Span)
         {
             const Real EPSILON = 0.01f;
